refactor: Switch to a multi-version build system
This allows us to support 1.18 and 1.19 at the same time, the reason for
this change is simple: calling methods from interfaces and from classes
is different in the bytecode.

On 1.19, the `Text` class is just a normal class with a method called
`copyContentOnly`. On 1.18, `Text` is an interface with the method name
`shallowCopy`. These are incompatible, even if they get remapped, as the
bytecode for calling a method from an interface and a method from a
class is different.
- Calling the method in 1.19: ``INVOKEVIRTUAL ...``
- Calling the method in 1.18: ``INVOKEINTERFACE ...``

Using a multi-version build project allows us to compile 2 seperate JARs
while still having one codebase. The replaymod preprocessor has been
added in case we need to conditionally add code in the future.