refactor: Improve how we store compacted messages
Previously, we were storing compacted messages in a hashmap of the
message content, as a string, to the message counter, as an integer.
This was fine, but it could've been better.

A class called `CompactedMessage` now exists which is a lot more capable
than just storing String -> int. Firstly, we can store the original
message and occurrences in the same object, meaning we can have some
nice functions inside this object like `getOriginalText` and
`getModifiedText` to make the detection of messages to compact easier.