ChatMessagesMixin: Cache the result of `breakRenderedChatMessageLines`
This is an extremely expensive operation which is run each time we
increment occurrences, so caching it is desirable in certain situations.
This cache is cleared whenever the chat is reset (e.g. when leaving
the world, or when pressing the key-bind).

This isn't as important in vanilla, as they don't refresh the message
list as often as we do.

This, along with the previous commit, prevents lag when a ton of long
messages with the same content are being spammed.

Fixes: #27